refactor(db): centralise les colonnes partagées
