+++ packages/db/src/domains/player/schema.ts
@@ -1,8 +1,8 @@
 import { sql } from 'drizzle-orm';
 import { pgTable, serial, varchar, bigint, integer, boolean } from 'drizzle-orm/pg-core';
 
+import { timestamps } from '../../shared/columns/timestampColumns.js';
 import { MAX_CHARACTER_NAME_LENGTH, MAX_CREW_NAME_LENGTH } from '../../shared/constants.js';
-import { timestamps } from '../../shared/helpers.js';
 import { guild } from '../guild/schema.js';
 import { subZoneEnum } from '../navigation/sub-zone-enum.js';
 import { ISLAND_ENTRY_SUB_ZONE, type Island } from '../navigation/world/registry.js';
