+++ packages/db/src/domains/resource/resource_instance/schema.ts
@@ -1,7 +1,7 @@
 import { sql } from 'drizzle-orm';
 import { check, integer, pgTable, serial, uniqueIndex } from 'drizzle-orm/pg-core';
 
-import { timestamps } from '../../../shared/helpers.js';
+import { timestamps } from '../../../shared/columns/timestampColumns.js';
 import { player } from '../../player/schema.js';
 import { resourceTemplate } from '../resource_template/schema.js';
