+++ packages/db/src/domains/resource/resource_template/schema.ts
@@ -1,8 +1,9 @@
 import { sql } from 'drizzle-orm';
 import { index, pgTable, serial, text, varchar } from 'drizzle-orm/pg-core';
 
-import { imageUrl, timestamps } from '../../../shared/helpers.js';
-import { rarityColumn } from '../../../shared/rarity-enum.js';
+import { imageUrl } from '../../../shared/columns/imageUrlColumn.js';
+import { rarityColumn } from '../../../shared/columns/rarityColumn.js';
+import { timestamps } from '../../../shared/columns/timestampColumns.js';
 
 import type { ResourceName } from './data.js';
 
@@ -11,7 +12,7 @@ export const resourceTemplate = pgTable(
   {
     id: serial('id').primaryKey(),
     name: varchar('name', { length: 128 }).$type<ResourceName>().notNull().unique(),
-    ...rarityColumn(),
+    rarity: rarityColumn(),
 
     ...imageUrl(),
     description: text('description'),
