+++ packages/db/src/domains/ship/schema.ts
@@ -1,6 +1,6 @@
 import { integer, pgTable, serial, varchar } from 'drizzle-orm/pg-core';
 
-import { timestamps } from '../../shared/helpers.js';
+import { timestamps } from '../../shared/columns/timestampColumns.js';
 import { player } from '../player/schema.js';
 
 export const ship = pgTable('ship', {
