+++ packages/db/src/schema.ts
@@ -7,7 +7,7 @@ export { onboardingStepEnum, player } from './domains/player/index.js';
 export { ship } from './domains/ship/index.js';
 export { resourceInstance, resourceTemplate } from './domains/resource/index.js';
 export { characterInstance, characterTemplate } from './domains/character/index.js';
-export { rarityEnum } from './shared/rarity-enum.js';
-export { pokemonType } from './shared/pokemon-type-enum.js';
+export { rarity as rarityEnum } from './shared/enums/rarity.js';
+export { pokemonType } from './shared/enums/pokemon-type.js';
 export { characterRaceEnum } from './domains/character/enum.js';
 export { characterSkillEnum } from './domains/character/skill-enum.js';
