+++ packages/db/src/shared/columns/imageUrlColumn.ts
@@ -1,0 +1,8 @@
+import { text } from 'drizzle-orm/pg-core';
+
+export function imageUrl({ isNullable = true }: { isNullable?: boolean } = {}) {
+  const column = text('image_url');
+  return {
+    imageUrl: isNullable ? column : column.notNull(),
+  };
+}
