+++ packages/db/src/shared/columns/pokemonTypesColumn.ts
@@ -1,0 +1,10 @@
+import { sql } from 'drizzle-orm';
+
+import { pokemonType } from '../enums/pokemon-type.js';
+
+export function pokemonTypes() {
+  return pokemonType('types')
+    .array()
+    .notNull()
+    .default(sql`'{}'::pokemon_type[]`);
+}
