+++ packages/db/src/shared/columns/rarityColumn.ts
@@ -1,0 +1,5 @@
+import { rarity } from '../enums/rarity.js';
+
+export function rarityColumn() {
+  return rarity('rarity').notNull().default('D');
+}
