+++ packages/db/src/shared/columns/timestampColumns.ts
@@ -1,4 +1,4 @@
-import { text, timestamp } from 'drizzle-orm/pg-core';
+import { timestamp } from 'drizzle-orm/pg-core';
 
 export function timestamps() {
   return {
@@ -7,12 +7,5 @@ export function timestamps() {
       .defaultNow()
       .notNull()
       .$onUpdate(() => new Date()),
-  };
-}
-
-export function imageUrl({ isNullable = true }: { isNullable?: boolean } = {}) {
-  const column = text('image_url');
-  return {
-    imageUrl: isNullable ? column : column.notNull(),
   };
 }
