+++ packages/db/src/shared/enums/pokemon-type.ts
