+++ packages/db/src/shared/enums/rarity.ts
@@ -1,0 +1,3 @@
+import { pgEnum } from 'drizzle-orm/pg-core';
+
+export const rarity = pgEnum('rarity', ['D', 'C', 'B', 'A', 'S', 'SS', 'X']);
